+++ prisma/migrations/20260731213146_add_metrics_url/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "apps" ADD COLUMN     "metrics_url" VARCHAR(500);